requirements.txt
agate==1.9.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
bandit==1.9.4
certifi==2026.2.25
cfgv==3.5.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
daff==1.4.2
-e git+https://github.com/evandycke/data-forge.git@a5d56b00faa4a7a0a9f915ce220a4c092a8a9700#egg=data_forge
dbt-adapters==1.22.8
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
detect-secrets==1.5.0
diff_cover==10.2.0
distlib==0.4.0
dlt==1.23.0
et_xmlfile==2.0.0
fastapi==0.135.1
filelock==3.25.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
giturlparse==0.14.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.15.0
identify==2.6.17
idna==3.11
importlib_metadata==8.7.1
inflect==7.5.0
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jinja2-simple-tags==0.6.1
jsonpath-ng==1.7.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
librt==0.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdformat==0.7.22
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.2
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orderly-set==5.5.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.2.0
platformdirs==4.9.4
pluggy==1.6.0
ply==3.11
pre_commit==4.5.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-discovery==1.1.1
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requirements-parser==0.13.0
rich==14.3.3
rich-argparse==1.7.2
rpds-py==0.30.0
ruamel.yaml==0.17.40
ruff==0.15.5
semver==3.0.4
setuptools==82.0.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
soda-core==3.5.6
soda-core-postgres==3.5.6
SQLAlchemy==2.0.48
sqlfluff==4.0.4
sqlfluff-templater-dbt==4.0.4
sqlglot==29.0.1
sqlparse==0.5.4
starlette==0.52.1
stevedore==5.7.0
tblib==3.2.2
tenacity==9.1.4
text-unidecode==1.3
tomlkit==0.14.0
tqdm==4.67.3
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.1.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
